requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-celerybeat-status==0.0.10
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-redis==4.11.0
django-redis-cache==2.1.1
django-rest-auth==0.9.5
django-shell-plus==1.1.7
djangorestframework==3.10.3
flower==0.9.4
gunicorn==20.0.4
humanize==0.5.1
idna==2.8
importlib-metadata==1.6.0
isodate==0.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.4.2
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.1.0
phonenumberslite==8.11.1
Pillow==7.1.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyhubtel-sms==0.0.1
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
redis==3.5.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
sentry-sdk==0.14.3
six==1.13.0
sqlparse==0.3.0
tornado==6.0.4
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
wrapt==1.11.2
zeep==3.4.0
zipp==3.1.0
